requirements.txt
Flask==1.1.2
flask_mail==0.9.1
PyYAML==5.4.1
